split out some features for postgres-util
