+++ src/postgres-util/src/lib.rs
@@ -75,28 +75,37 @@
 
 //! PostgreSQL utility library.
 
-use std::time::Duration;
-
-use desc::PostgresSchemaDesc;
 use openssl::pkey::PKey;
 use openssl::ssl::{SslConnector, SslMethod, SslVerifyMode};
 use openssl::x509::X509;
 use postgres_openssl::MakeTlsConnector;
-use serde::{Deserialize, Serialize};
-use tokio::net::TcpStream as TokioTcpStream;
-use tokio_postgres::config::{Host, ReplicationMode, SslMode};
-use tokio_postgres::tls::MakeTlsConnect;
-use tokio_postgres::types::Oid;
-use tokio_postgres::Client;
+use tokio_postgres::config::SslMode;
 use tracing::warn;
 
-use mz_ore::task;
-use mz_repr::GlobalId;
-use mz_ssh_util::tunnel::SshTunnelConfig;
+macro_rules! bail_generic {
+    ($fmt:expr, $($arg:tt)*) => {
+        return Err(PostgresError::Generic(anyhow::anyhow!($fmt, $($arg)*)))
+    };
+    ($err:expr $(,)?) => {
+        return Err(PostgresError::Generic(anyhow::anyhow!($err)))
+    };
+}
 
-use crate::desc::{PostgresColumnDesc, PostgresKeyDesc, PostgresTableDesc};
-
+#[cfg(feature = "schemas")]
 pub mod desc;
+#[cfg(feature = "schemas")]
+pub mod schemas;
+#[cfg(feature = "schemas")]
+pub use schemas::{get_schemas, publication_info};
+#[cfg(feature = "tunnel")]
+pub mod tunnel;
+#[cfg(feature = "tunnel")]
+pub use tunnel::{
+    drop_replication_slots, Config, ReplicationTimeouts, TunnelConfig,
+    DEFAULT_REPLICATION_CONNECT_TIMEOUT, DEFAULT_REPLICATION_KEEPALIVE_IDLE,
+    DEFAULT_REPLICATION_KEEPALIVE_INTERVAL, DEFAULT_REPLICATION_KEEPALIVE_RETRIES,
+    DEFAULT_REPLICATION_TCP_USER_TIMEOUT,
+};
 
 /// An error representing pg, ssh, ssl, and other failures.
 #[derive(Debug, thiserror::Error)]
@@ -105,6 +114,7 @@ pub enum PostgresError {
     #[error(transparent)]
     Generic(#[from] anyhow::Error),
     /// Error using ssh.
+    #[cfg(feature = "tunnel")]
     #[error(transparent)]
     Ssh(#[from] openssh::Error),
     /// Error doing io to setup an ssh connection.
@@ -116,15 +126,6 @@ pub enum PostgresError {
     /// Error setting up postgres ssl.
     #[error(transparent)]
     PostgresSsl(#[from] openssl::error::ErrorStack),
-}
-
-macro_rules! bail_generic {
-    ($fmt:expr, $($arg:tt)*) => {
-        return Err(PostgresError::Generic(anyhow::anyhow!($fmt, $($arg)*)))
-    };
-    ($err:expr $(,)?) => {
-        return Err(PostgresError::Generic(anyhow::anyhow!($err)))
-    };
 }
 
 /// Creates a TLS connector for the given [`Config`].
@@ -186,382 +187,4 @@ pub fn make_tls(config: &tokio_postgres::Config) -> Result<MakeTlsConnector, Pos
     }
 
     Ok(tls_connector)
-}
-
-pub async fn get_schemas(config: &Config) -> Result<Vec<PostgresSchemaDesc>, PostgresError> {
-    let client = config.connect("postgres_schemas").await?;
-
-    Ok(client
-        .query("SELECT oid, nspname, nspowner FROM pg_namespace", &[])
-        .await?
-        .into_iter()
-        .map(|row| {
-            let oid: Oid = row.get("oid");
-            let name: String = row.get("nspname");
-            let owner: Oid = row.get("nspowner");
-            PostgresSchemaDesc { oid, name, owner }
-        })
-        .collect::<Vec<_>>())
-}
-
-/// Fetches table schema information from an upstream Postgres source for
-/// tables that are part of a publication, given a connection string and the
-/// publication name.
-///
-/// If `oid_filter` is `None`, returns all tables, otherwise returns only the
-/// details for the identified oid.
-///
-/// # Errors
-///
-/// - Invalid connection string, user information, or user permissions.
-/// - Upstream publication does not exist or contains invalid values.
-pub async fn publication_info(
-    config: &Config,
-    publication: &str,
-    oid_filter: Option<u32>,
-) -> Result<Vec<PostgresTableDesc>, PostgresError> {
-    let client = config.connect("postgres_publication_info").await?;
-
-    client
-        .query(
-            "SELECT oid FROM pg_publication WHERE pubname = $1",
-            &[&publication],
-        )
-        .await?
-        .get(0)
-        .ok_or_else(|| anyhow::anyhow!("publication {:?} does not exist", publication))?;
-
-    let tables = client
-        .query(
-            "SELECT
-                c.oid, p.schemaname, p.tablename
-            FROM
-                pg_catalog.pg_class AS c
-                JOIN pg_namespace AS n ON c.relnamespace = n.oid
-                JOIN pg_publication_tables AS p ON
-                        c.relname = p.tablename AND n.nspname = p.schemaname
-            WHERE
-                p.pubname = $1
-                AND ($2::oid IS NULL OR c.oid = $2::oid)",
-            &[&publication, &oid_filter],
-        )
-        .await?;
-
-    let mut table_infos = vec![];
-    for row in tables {
-        let oid = row.get("oid");
-
-        let columns = client
-            .query(
-                "SELECT
-                        a.attname AS name,
-                        a.atttypid AS typoid,
-                        a.attnum AS colnum,
-                        a.atttypmod AS typmod,
-                        a.attnotnull AS not_null,
-                        b.oid IS NOT NULL AS primary_key
-                    FROM pg_catalog.pg_attribute a
-                    LEFT JOIN pg_catalog.pg_constraint b
-                        ON a.attrelid = b.conrelid
-                        AND b.contype = 'p'
-                        AND a.attnum = ANY (b.conkey)
-                    WHERE a.attnum > 0::pg_catalog.int2
-                        AND NOT a.attisdropped
-                        AND a.attrelid = $1
-                    ORDER BY a.attnum",
-                &[&oid],
-            )
-            .await?
-            .into_iter()
-            .map(|row| {
-                let name: String = row.get("name");
-                let type_oid = row.get("typoid");
-                let col_num = Some(
-                    row.get::<_, i16>("colnum")
-                        .try_into()
-                        .expect("non-negative values"),
-                );
-                let type_mod: i32 = row.get("typmod");
-                let not_null: bool = row.get("not_null");
-                Ok(PostgresColumnDesc {
-                    name,
-                    col_num,
-                    type_oid,
-                    type_mod,
-                    nullable: !not_null,
-                })
-            })
-            .collect::<Result<Vec<_>, PostgresError>>()?;
-
-        // PG 15 adds UNIQUE NULLS NOT DISTINCT, which would let us use `UNIQUE` constraints over
-        // nullable columns as keys; i.e. aligns a PG index's NULL handling with an arrangement's
-        // keys. For more info, see https://www.postgresql.org/about/featurematrix/detail/392/
-        let pg_15_plus_keys = "
-        SELECT
-            pg_constraint.oid,
-            pg_constraint.conkey,
-            pg_constraint.conname,
-            pg_constraint.contype = 'p' AS is_primary,
-            pg_index.indnullsnotdistinct AS nulls_not_distinct
-        FROM
-            pg_constraint
-                JOIN
-                    pg_index
-                    ON pg_index.indexrelid = pg_constraint.conindid
-        WHERE
-            pg_constraint.conrelid = $1
-                AND
-            pg_constraint.contype =ANY (ARRAY['p', 'u']);";
-
-        // As above but for versions of PG without indnullsnotdistinct.
-        let pg_14_minus_keys = "
-        SELECT
-            pg_constraint.oid,
-            pg_constraint.conkey,
-            pg_constraint.conname,
-            pg_constraint.contype = 'p' AS is_primary,
-            false AS nulls_not_distinct
-        FROM pg_constraint
-        WHERE
-            pg_constraint.conrelid = $1
-                AND
-            pg_constraint.contype =ANY (ARRAY['p', 'u']);";
-
-        let keys = match client.query(pg_15_plus_keys, &[&oid]).await {
-            Ok(keys) => keys,
-            Err(e)
-                // PG versions prior to 15 do not contain this column.
-                if e.to_string()
-                    == "db error: ERROR: column pg_index.indnullsnotdistinct does not exist" =>
-            {
-                client.query(pg_14_minus_keys, &[&oid]).await?
-            }
-            e => e?,
-        };
-
-        let keys = keys
-            .into_iter()
-            .map(|row| {
-                let oid: u32 = row.get("oid");
-                let cols: Vec<i16> = row.get("conkey");
-                let name: String = row.get("conname");
-                let is_primary: bool = row.get("is_primary");
-                let nulls_not_distinct: bool = row.get("nulls_not_distinct");
-                let cols = cols
-                    .into_iter()
-                    .map(|col| u16::try_from(col).expect("non-negative colnums"))
-                    .collect();
-                PostgresKeyDesc {
-                    oid,
-                    name,
-                    cols,
-                    is_primary,
-                    nulls_not_distinct,
-                }
-            })
-            .collect();
-
-        table_infos.push(PostgresTableDesc {
-            oid,
-            namespace: row.get("schemaname"),
-            name: row.get("tablename"),
-            columns,
-            keys,
-        });
-    }
-
-    Ok(table_infos)
-}
-
-pub async fn drop_replication_slots(config: Config, slots: &[&str]) -> Result<(), PostgresError> {
-    let client = config.connect("postgres_drop_replication_slots").await?;
-    let replication_client = config.connect_replication().await?;
-    for slot in slots {
-        let rows = client
-            .query(
-                "SELECT active_pid FROM pg_replication_slots WHERE slot_name = $1::TEXT",
-                &[&slot],
-            )
-            .await?;
-        match rows.len() {
-            0 => {
-                // DROP_REPLICATION_SLOT will error if the slot does not exist
-                // todo@jldlaughlin: don't let invalid Postgres sources ship!
-                continue;
-            }
-            1 => {
-                replication_client
-                    .simple_query(&format!("DROP_REPLICATION_SLOT {} WAIT", slot))
-                    .await?;
-            }
-            _ => {
-                return Err(PostgresError::Generic(anyhow::anyhow!(
-                    "multiple pg_replication_slots entries for slot {}",
-                    &slot
-                )))
-            }
-        }
-    }
-    Ok(())
-}
-
-/// Configures an optional tunnel for use when connecting to a PostgreSQL
-/// database.
-#[derive(Debug, PartialEq, Clone)]
-pub enum TunnelConfig {
-    /// Establish a direct TCP connection to the database host.
-    Direct,
-    /// Establish a TCP connection to the database via an SSH tunnel.
-    /// This means first establishing an SSH connection to a bastion host,
-    /// and then opening a separate connection from that host to the database.
-    /// This is commonly referred by vendors as a "direct SSH tunnel", in
-    /// opposition to "reverse SSH tunnel", which is currently unsupported.
-    Ssh(SshTunnelConfig),
-    /// Establish a TCP connection to the database via an AWS PrivateLink
-    /// service.
-    AwsPrivatelink {
-        /// The ID of the AWS PrivateLink service.
-        connection_id: GlobalId,
-    },
-}
-
-/// Configurable timeouts that apply only when using [`Config::connect_replication`].
-#[derive(Clone, Debug, Default, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ReplicationTimeouts {
-    pub connect_timeout: Option<Duration>,
-    pub keepalives_retries: Option<u32>,
-    pub keepalives_idle: Option<Duration>,
-    pub keepalives_interval: Option<Duration>,
-    pub tcp_user_timeout: Option<Duration>,
-}
-
-// Some of these defaults were recommended by @ph14
-// https://github.com/MaterializeInc/materialize/pull/18644#discussion_r1160071692
-pub const DEFAULT_REPLICATION_CONNECT_TIMEOUT: Duration = Duration::from_secs(30);
-pub const DEFAULT_REPLICATION_KEEPALIVE_INTERVAL: Duration = Duration::from_secs(10);
-pub const DEFAULT_REPLICATION_KEEPALIVE_IDLE: Duration = Duration::from_secs(10);
-pub const DEFAULT_REPLICATION_KEEPALIVE_RETRIES: u32 = 5;
-// This is meant to be DEFAULT_REPLICATION_KEEPALIVE_IDLE
-// + DEFAULT_REPLICATION_KEEPALIVE_RETRIES * DEFAULT_REPLICATION_KEEPALIVE_INTERVAL
-pub const DEFAULT_REPLICATION_TCP_USER_TIMEOUT: Duration = Duration::from_secs(60);
-
-/// Configuration for PostgreSQL connections.
-///
-/// This wraps [`tokio_postgres::Config`] to allow the configuration of a
-/// tunnel via a [`TunnelConfig`].
-#[derive(Debug, PartialEq, Clone)]
-pub struct Config {
-    inner: tokio_postgres::Config,
-    tunnel: TunnelConfig,
-    replication_timeouts: ReplicationTimeouts,
-}
-
-impl Config {
-    pub fn new(inner: tokio_postgres::Config, tunnel: TunnelConfig) -> Result<Self, PostgresError> {
-        let config = Self {
-            inner,
-            tunnel,
-            replication_timeouts: ReplicationTimeouts::default(),
-        };
-
-        // Early validate that the configuration contains only a single TCP
-        // server.
-        config.address()?;
-
-        Ok(config)
-    }
-
-    pub fn replication_timeouts(mut self, replication_timeouts: ReplicationTimeouts) -> Config {
-        self.replication_timeouts = replication_timeouts;
-        self
-    }
-
-    /// Connects to the configured PostgreSQL database.
-    pub async fn connect(&self, task_name: &str) -> Result<Client, PostgresError> {
-        self.connect_internal(task_name, |_| ()).await
-    }
-
-    /// Starts a replication connection to the configured PostgreSQL database.
-    pub async fn connect_replication(&self) -> Result<Client, PostgresError> {
-        self.connect_internal("postgres_connect_replication", |config| {
-            config
-                .replication_mode(ReplicationMode::Logical)
-                .connect_timeout(
-                    self.replication_timeouts
-                        .connect_timeout
-                        .unwrap_or(DEFAULT_REPLICATION_CONNECT_TIMEOUT),
-                )
-                .keepalives_interval(
-                    self.replication_timeouts
-                        .keepalives_interval
-                        .unwrap_or(DEFAULT_REPLICATION_KEEPALIVE_INTERVAL),
-                )
-                .keepalives_idle(
-                    self.replication_timeouts
-                        .keepalives_idle
-                        .unwrap_or(DEFAULT_REPLICATION_KEEPALIVE_IDLE),
-                )
-                .keepalives_retries(
-                    self.replication_timeouts
-                        .keepalives_retries
-                        .unwrap_or(DEFAULT_REPLICATION_KEEPALIVE_RETRIES),
-                )
-                .tcp_user_timeout(
-                    self.replication_timeouts
-                        .tcp_user_timeout
-                        .unwrap_or(DEFAULT_REPLICATION_TCP_USER_TIMEOUT),
-                );
-        })
-        .await
-    }
-
-    fn address(&self) -> Result<(&str, u16), PostgresError> {
-        match (self.inner.get_hosts(), self.inner.get_ports()) {
-            ([Host::Tcp(host)], [port]) => Ok((host, *port)),
-            _ => bail_generic!("only TCP connections to a single PostgreSQL server are supported"),
-        }
-    }
-
-    async fn connect_internal<F>(
-        &self,
-        task_name: &str,
-        configure: F,
-    ) -> Result<Client, PostgresError>
-    where
-        F: FnOnce(&mut tokio_postgres::Config),
-    {
-        let mut postgres_config = self.inner.clone();
-        configure(&mut postgres_config);
-        let mut tls = make_tls(&postgres_config)?;
-        match &self.tunnel {
-            TunnelConfig::Direct => {
-                let (client, connection) = postgres_config.connect(tls).await?;
-                task::spawn(|| task_name, connection);
-                Ok(client)
-            }
-            TunnelConfig::Ssh(tunnel) => {
-                let (host, port) = self.address()?;
-                let tunnel = tunnel.connect(host, port).await?;
-                let tls = MakeTlsConnect::<TokioTcpStream>::make_tls_connect(&mut tls, host)?;
-                let tcp_stream = TokioTcpStream::connect(tunnel.local_addr()).await?;
-                let (client, connection) = postgres_config.connect_raw(tcp_stream, tls).await?;
-                task::spawn(|| task_name, async {
-                    let _tunnel = tunnel; // Keep SSH tunnel alive for duration of connection.
-                    if let Err(e) = connection.await {
-                        warn!("postgres connection failed: {e}");
-                    }
-                });
-                Ok(client)
-            }
-            TunnelConfig::AwsPrivatelink { connection_id } => {
-                let (host, port) = self.address()?;
-                let privatelink_host = mz_cloud_resources::vpc_endpoint_name(*connection_id);
-                let tls = MakeTlsConnect::<TokioTcpStream>::make_tls_connect(&mut tls, host)?;
-                let tcp_stream = TokioTcpStream::connect((privatelink_host, port)).await?;
-                let (client, connection) = postgres_config.connect_raw(tcp_stream, tls).await?;
-                task::spawn(|| task_name, connection);
-                Ok(client)
-            }
-        }
-    }
 }
